+++ src/test/java/com/project/zipkok/user/fixture/AuthTokensFixture.java
@@ -1,4 +1,4 @@
-package com.project.zipkok.fixture;
+package com.project.zipkok.user.fixture;
 
 import com.project.zipkok.util.jwt.AuthTokens;
