+++ src/test/java/com/project/zipkok/user/fixture/JwtUserDetailsFixture.java
@@ -1,4 +1,4 @@
-package com.project.zipkok.fixture;
+package com.project.zipkok.user.fixture;
 
 import com.project.zipkok.common.enums.Role;
 import com.project.zipkok.util.jwt.JwtUserDetails;
