+++ src/test/java/com/project/zipkok/user/infrastructure/UserRepositoryTest.java
@@ -1,7 +1,8 @@
-package com.project.zipkok.repository;
+package com.project.zipkok.user.infrastructure;
 
-import com.project.zipkok.fixture.*;
 import com.project.zipkok.model.*;
+import com.project.zipkok.repository.*;
+import com.project.zipkok.user.fixture.*;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -12,7 +13,6 @@ import org.springframework.test.context.TestPropertySource;
 
 import java.util.Optional;
 
-import static org.mockito.BDDMockito.given;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.*;
